requirements.txt
Django==1.9.7
Pillow==3.2.0
cssmin==0.2.0
django-appconf==1.0.2
django-autoslug==1.9.3
django-bower==5.1.0
django-compressor==1.6
django-environ==0.3.0
django-extensions==1.6.7
djangorestframework==3.3.3
djangorestframework-gis==0.10.1
geojson==1.3.2
gunicorn==19.6.0
ipython==4.2.0
numpy==1.11.0
psycopg2==2.6.1
pyproj==1.9.5.1
pyshp==1.2.3
six==1.10.0
versiontools==1.9.1
virtual-node==0.1.0
wsgiref==0.1.2
